requirements.txt
numpy>=1.24
pandas>=1.5
scikit-learn>=1.2
lightgbm>=3.3
joblib>=1.2
matplotlib>=3.6
seaborn>=0.12
tensorflow-cpu>=2.12
scipy>=1.9
